abstract common components for batch/streaming metrics
