+++ scio-test/src/it/scala/com/spotify/ScioStreamingBenchmark.scala
@@ -17,47 +17,20 @@
 
 package com.spotify
 
-import java.util.concurrent.{ScheduledThreadPoolExecutor, TimeUnit}
-
-import com.google.api.client.googleapis.auth.oauth2.GoogleCredential
-import com.google.api.client.googleapis.javanet.GoogleNetHttpTransport
-import com.google.api.client.json.jackson2.JacksonFactory
-import com.google.api.services.dataflow.Dataflow
 import com.google.api.services.dataflow.model.Job
 import com.google.common.reflect.ClassPath
 import com.spotify.scio._
-import com.spotify.scio.runners.dataflow.DataflowResult
 import org.apache.beam.sdk.io.GenerateSequence
 import org.apache.beam.sdk.options.StreamingOptions
 import org.joda.time._
 import org.joda.time.format.DateTimeFormat
 
 import scala.collection.JavaConverters._
-import scala.concurrent.duration.Duration
 
-
+/** Streaming benchmark jobs, restarted daily and polled for metrics every hour. */
 object ScioStreamingBenchmark {
-
-  object Settings {
-    val defaultProjectId: String = "data-integration-test"
-    val numOfWorkers = 4
-    val commonArgs = Array(
-      "--runner=DataflowRunner",
-      s"--numWorkers=$numOfWorkers",
-      "--workerMachineType=n1-highmem-8",
-      "--autoscalingAlgorithm=NONE")
-
-    val shuffleConf = Map("ShuffleService" -> Array("--experiments=shuffle_mode=service"))
-
-    val dataflow: Dataflow = {
-      val transport = GoogleNetHttpTransport.newTrustedTransport()
-      val jackson = JacksonFactory.getDefaultInstance
-      val credential = GoogleCredential.getApplicationDefault
-      new Dataflow.Builder(transport, jackson, credential).build()
-    }
-  }
-
-  private val executor = new ScheduledThreadPoolExecutor(1)
+  import DataflowProvider._
+  import ScioBenchmarkSettings._
 
   private val benchmarks = ClassPath.from(Thread.currentThread().getContextClassLoader)
     .getAllClasses
@@ -76,7 +49,7 @@ object ScioStreamingBenchmark {
     val argz = Args(args)
     val name = argz("name")
     val regex = argz.getOrElse("regex", ".*")
-    val projectId = argz.getOrElse("project", Settings.defaultProjectId)
+    val projectId = argz.getOrElse("project", defaultProjectId)
     val timestamp = DateTimeFormat.forPattern("yyyyMMddHHmmss")
       .withZone(DateTimeZone.UTC)
       .print(System.currentTimeMillis())
@@ -84,24 +57,16 @@ object ScioStreamingBenchmark {
 
     cancelCurrentJobs(projectId)
 
-    val scioResults = benchmarks
+    benchmarks
       .filter(_.name.matches(regex))
-      .map(_.run(projectId, prefix, Settings.commonArgs))
-
-    executor.scheduleAtFixedRate(
-      () => scioResults.foreach { case (jobName, result) => pollMetrics(jobName, result) },
-      1,
-      60,
-      TimeUnit.MINUTES)
-
-    scioResults.foreach(result => result._2.waitUntilFinish(Duration(1, TimeUnit.DAYS)))
+      .map(_.run(projectId, prefix, commonArgs("n1-highmem-8")))
   }
 
   private val cancelledState = "JOB_STATE_CANCELLED"
 
   /** Cancel any currently running streaming benchmarks before spawning new ones */
   private def cancelCurrentJobs(projectId: String): Unit = {
-    val jobs = Settings.dataflow.projects().jobs()
+    val jobs = dataflow.projects().jobs()
 
     Option(jobs.list(projectId).setFilter("ACTIVE").execute().getJobs).foreach { activeJobs =>
       activeJobs.asScala.foreach { job =>
@@ -115,25 +80,6 @@ object ScioStreamingBenchmark {
           ).execute()
         }
       }
-    }
-  }
-
-  // @Todo write to DataStore etc
-  private def pollMetrics(benchmark: String, result: ScioResult): Unit = {
-    PrettyPrint.print("PollMetrics", s"polling metrics for " + benchmark)
-
-    try {
-      result.as[DataflowResult]
-        .getJobMetrics.getMetrics.asScala
-        .filter { metric =>
-          val name = metric.getName.getName
-          name.startsWith("Total") || name.startsWith("Current")
-        }.foreach { metric =>
-          PrettyPrint.print(benchmark, s"${metric.getName.getName} -> ${metric.getScalar}")
-        }
-    } catch {
-      case e: Exception =>
-        PrettyPrint.print("PollMetrics", s"caught exception polling for metrics: $e")
     }
   }
 
@@ -161,6 +107,46 @@ object ScioStreamingBenchmark {
   }
 }
 
+/** Triggered once an hour to read job metrics from currently running streaming benchmarks
+ * and write to DataStore. */
+object ScioStreamingBenchmarkMetrics {
+  import DataflowProvider._
+  import ScioBenchmarkSettings._
+
+  def main(args: Array[String]): Unit = {
+    val argz = Args(args)
+    val name = argz("name")
+    val projectId = argz.getOrElse("project", defaultProjectId)
+    val jobNamePattern = s"sciostreamingbenchmark-$name-\\d+-([a-zA-Z0-9]+)-\\w+".r
+
+    val jobs = dataflow.projects().jobs().list(projectId)
+
+    val hourlyMetrics = Option(jobs.setFilter("ACTIVE").execute().getJobs)
+      .map { activeJobs =>
+        activeJobs.asScala.flatMap { job =>
+          if (job.getName.toLowerCase.startsWith("sciostreamingbenchmark")) {
+            jobNamePattern.findFirstMatchIn(job.getName).map(_.group(1))
+              .map { benchmarkName =>
+                BenchmarkResult.streaming(
+                  benchmarkName,
+                  job.getCreateTime,
+                  dataflow.projects().jobs().getMetrics(projectId, job.getId).execute())
+              }.orElse {
+              PrettyPrint.print(
+                "Active jobs fetcher",
+                s"Could not parse validbenchmark name from job ${job.getName}")
+              None
+            }
+          } else {
+            None
+          }
+        }.toList
+      }.getOrElse(List())
+
+    new DatastoreStreamingLogger().log(hourlyMetrics)
+  }
+}
+
 abstract class StreamingBenchmark {
   val name: String = this.getClass.getSimpleName.replaceAll("\\$$", "")
 
@@ -179,4 +165,15 @@ abstract class StreamingBenchmark {
   }
 
   def run(sc: ScioContext): Unit
+}
+
+class DatastoreStreamingLogger extends DatastoreLogger(ScioBenchmarkSettings.StreamingMetrics) {
+  override def dsKeyId(benchmark: BenchmarkResult, env: CircleCIEnv): String = {
+    s"${env.buildNum.toString}[${timeOffset(benchmark.startTime)}]"
+  }
+
+  private def timeOffset(startTime: LocalDateTime): String = {
+    val hourOffset = Hours.hoursBetween(startTime, new LocalDateTime(DateTimeZone.UTC)).getHours
+    s"+${"%02d".format(hourOffset)}h"
+  }
 }
